Refactor container
